Ensure pandas.errors only imports its __all__
